Experimental support of 64 bit SI type (through macro definition of SI_64 instead of SI_32)
git-svn-id: svn://svn.savannah.gnu.org/texmacs/trunk@14867 64cb5145-927a-446d-8aed-2fb7b4773692

+++ src/src/Plugins/Freetype/tt_analyze.cpp
@@ -236,7 +236,7 @@ max_ascent (font_metric fnm, int start, int end) {
   int m= 0;
   for (int i=start; i<=end; i++) {
     metric_struct* x= fnm->get (i);
-    m= max (m, x->y2/256);
+    m= max (m, (int) x->y2/256);
   }
   return m;
 }
@@ -246,7 +246,7 @@ max_descent (font_metric fnm, int start, int end) {
   int m= 0;
   for (int i=start; i<=end; i++) {
     metric_struct* x= fnm->get (i);
-    m= max (m, -x->y1/256);
+    m= max (m, (int) (-x->y1/256));
   }
   return m;
 }
